Fix TaskEmbeddingSampler parallelism
Moves the worker functions into global scope. They were originally class
methods on TaskEmbeddingSampler. The argument list provided to
multiprocessing was missing `self`. We could provide `self`, but there's no
need because the workers use only global and local scopes.

Note: the design of singleton_pool.run_each() assumes that target worker
method is not bound to a class. This should at least be guarded with an
`assert`